nix: Use mayastor-overlay to build liburing (#121)
Upstream nix-unstable moved from liburing 0.2 to 0.4pre, which introduced libdevdir in the configure script. This was not set as part of the build so libspdk_fat.so ended up being linked to liburing.so.1 without a full path, breaking the build of Mayastor.

Override upstream's nix expression by setting libdevdir to libdir, and also moving up to the current 0.5 release.